Split up AST.Traversable1 into two modules.
